tests/mknod.c: print syscall() result using %ld format

+++ tests/mknod.c
@@ -20,17 +20,16 @@
 int
 main(void)
 {
-	int rc = syscall(__NR_mknod, TMP_FILE, S_IFREG|0600, 0);
-	printf("mknod(\"%s\", S_IFREG|0600) = %d %s (%m)\n",
-	       TMP_FILE, rc,
-	       errno2name());
+	long rc = syscall(__NR_mknod, TMP_FILE, S_IFREG|0600, 0);
+	printf("mknod(\"%s\", S_IFREG|0600) = %ld %s (%m)\n",
+	       TMP_FILE, rc, errno2name());
 
 	const unsigned long dev =
 		(unsigned long) 0xdeadbeef00000000 | makedev(1, 7);
+
 	rc = syscall(__NR_mknod, TMP_FILE, S_IFCHR | 0400, dev);
-	printf("mknod(\"%s\", S_IFCHR|0400, makedev(1, 7)) = %d %s (%m)\n",
-	       TMP_FILE, rc,
-	       errno2name());
+	printf("mknod(\"%s\", S_IFCHR|0400, makedev(1, 7)) = %ld %s (%m)\n",
+	       TMP_FILE, rc, errno2name());
 
 	puts("+++ exited with 0 +++");
 	return 0;
